move compare_strings to utils.py
